+++ src/network/mocks.ts
@@ -1,0 +1,8 @@
+import { NetworkAdapter } from "./types";
+
+export const networkMock: NetworkAdapter<any> = function() {
+  return {
+    getAndObserve: () => () => {},
+    push: () => {}
+  };
+};
